FIX: layout for all routes
