+++ app/(routes)/layout.tsx
@@ -1,11 +1,12 @@
 import type { Metadata } from 'next'
-import { ClerkProvider } from '@clerk/nextjs'
 
-import './globals.css'
+
+import '@/app/globals.css'
 import { Inter } from 'next/font/google'
-import { Main } from 'next/document'
-import MainFooter from './components/layout/footer'
-import Navigation from './utils/nav/navigation'
+
+import MainFooter from '../components/layout/footer'
+import Navigation from '../utils/nav/navigation'
+
 const inter = Inter({ subsets: ['latin'] })
 
 export const metadata: Metadata = {
@@ -13,21 +14,21 @@ export const metadata: Metadata = {
     description: 'Generated by create next app',
 }
 
-export default function RootLayout({
+export default function PublicLayout({
     children,
 }: Readonly<{
     children: React.ReactNode
 }>) {
     return (
-        <ClerkProvider>
+
             <html lang="en" suppressHydrationWarning>
                 <body className={inter.className}>
 
-                {/* <Navigation /> */}
+                <Navigation />
                     {children}
-                    {/* <MainFooter /> */}
+                    <MainFooter />
                 </body>
             </html>
-        </ClerkProvider>
+
     )
 }
